gpu: handle double loadScene()
This happens rarely if the map load is slow and also there are back to
back map build packets.

Reproduce via sleep(2000) in loadScene() and use Book of the dead to
teleport to Hosidius and run to the ladder south.

(cherry picked from commit 9cbcb9ed40929cb0a35ea6517b8f46963ea881ed)